refactor(BA-4300): Rebase on BA-4296 and use its operation context
